Init field within Field class
